+++ csatar-plugins/knowledgerepository/models/Region.php
@@ -29,6 +29,9 @@ class Region extends Model
 
     public $fillable = [
         'name',
+        'big_parent_id',
+        'mid_parent_id',
+        'small_parent_id'
     ];
 
     public $belongsTo = [
@@ -64,4 +67,24 @@ class Region extends Model
         ],
     ];
 
+    public function getExtendedNameAttribute()
+    {
+        $name = '';
+
+        if (!empty($this->big_parent_id)) {
+            $name.= $this->big_parent->name . ' - ';
+        }
+
+        if (!empty($this->mid_parent_id)) {
+            $name.= $this->mid_parent->name . ' - ';
+        }
+
+        if (!empty($this->small_parent_id)) {
+            $name.= $this->small_parent->name . ' - ';
+        }
+
+        $name.= $this->name;
+
+        return $name;
+    }
 }