+++ Changelog.md
@@ -16,6 +16,7 @@
 - CS-517 - Created Folk song type model
 - CS-518 - Created Region model
 - CS-519 - Created Folk song rhythm model
+- CS-508 - Created Song model
 
 ## 1.11.0
 ### 2023-04-13